+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupExportException.kt
@@ -1,5 +1,7 @@
 package com.thomaskioko.tvmaniac.data.backup.api
 
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupFailure
+
 public class BackupExportException(
     public val reason: BackupFailure,
     override val cause: Throwable? = null,
