--- data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupFailure.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.data.backup.api
-
-public enum class BackupFailure {
-    WriteFailed,
-    VerificationFailed,
-}
