+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupRepository.kt
@@ -1,5 +1,9 @@
 package com.thomaskioko.tvmaniac.data.backup.api
 
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupFile
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupResult
+import com.thomaskioko.tvmaniac.data.backup.api.model.RestoreResult
+
 public interface BackupRepository {
     public suspend fun createBackup(): BackupFile
