+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/RestoredListWriter.kt
@@ -1,5 +1,7 @@
 package com.thomaskioko.tvmaniac.data.backup.api
 
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupList
+
 public interface RestoredListWriter {
     public suspend fun restoreLists(lists: List<BackupList>): Int
 }
