+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/model/AutoBackupStatus.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.data.backup.api.model
+
+public data class AutoBackupStatus(
+    val lastRunAt: Long? = null,
+    val lastRunFailed: Boolean = false,
+)
