+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/model/BackupFailure.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.data.backup.api.model
+
+public enum class BackupFailure {
+    WriteFailed,
+    VerificationFailed,
+    LocationUnavailable,
+}
