+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/model/BackupResult.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.data.backup.api
+package com.thomaskioko.tvmaniac.data.backup.api.model
 
 public sealed interface BackupResult {
     public data class Success(val showCount: Int, val episodeCount: Int) : BackupResult
