+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/model/ShowRefillProgress.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.data.backup.api.model
+
+public data class ShowRefillProgress(
+    val completed: Int = 0,
+    val total: Int = 0,
+)
