+++ data/backup/implementation/src/commonTest/kotlin/com/thomaskioko/tvmaniac/data/backup/implementation/DefaultBackupRepositoryRestoreTest.kt
@@ -1,18 +1,18 @@
 package com.thomaskioko.tvmaniac.data.backup.implementation
 
 import com.thomaskioko.tvmaniac.core.base.model.AppCoroutineDispatchers
-import com.thomaskioko.tvmaniac.data.backup.api.BackupEpisode
-import com.thomaskioko.tvmaniac.data.backup.api.BackupFile
 import com.thomaskioko.tvmaniac.data.backup.api.BackupFormat
-import com.thomaskioko.tvmaniac.data.backup.api.BackupList
-import com.thomaskioko.tvmaniac.data.backup.api.BackupListShow
-import com.thomaskioko.tvmaniac.data.backup.api.BackupRating
-import com.thomaskioko.tvmaniac.data.backup.api.BackupSeason
-import com.thomaskioko.tvmaniac.data.backup.api.BackupSeasonRating
-import com.thomaskioko.tvmaniac.data.backup.api.BackupShow
-import com.thomaskioko.tvmaniac.data.backup.api.BackupWatchedEpisode
-import com.thomaskioko.tvmaniac.data.backup.api.RestoreFailure
-import com.thomaskioko.tvmaniac.data.backup.api.RestoreResult
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupEpisode
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupFile
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupList
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupListShow
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupRating
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupSeason
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupSeasonRating
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupShow
+import com.thomaskioko.tvmaniac.data.backup.api.model.BackupWatchedEpisode
+import com.thomaskioko.tvmaniac.data.backup.api.model.RestoreFailure
+import com.thomaskioko.tvmaniac.data.backup.api.model.RestoreResult
 import com.thomaskioko.tvmaniac.data.backup.testing.FakeBackupDestination
 import com.thomaskioko.tvmaniac.data.backup.testing.FakeRestoredListWriter
 import com.thomaskioko.tvmaniac.database.test.BaseDatabaseTest
